django-backend/concept_crafter/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
dj-database-url>=2.1
psycopg2-binary>=2.9
requests>=2.31
pytest>=7.4
pytest-django>=4.7
hypothesis>=6.88
factory-boy>=3.3
